fix: rename IAutoBeRealizeCorderApplication to IAutoBeRealizeCoderApplication and update related references
